refactor menu by use-navigation hook to generate the nav routes and subroutes
