+++ .vscode/settings.json
@@ -3,6 +3,8 @@
 		"Calipso",
 		"clsx",
 		"Competición",
+		"glassmorphism",
+		"headlessui",
 		"Khand",
 		"ossp",
 		"Peña",
