+++ src/app/globals.css
@@ -8,4 +8,8 @@
 		align-items: center;
 		justify-content: center;
 	}
+	/* glassmorphism background style */
+	.bg-glassmorphism {
+		@apply bg-gray-400 bg-opacity-10 bg-clip-padding backdrop-blur-sm backdrop-filter;
+	}
 }
