+++ src/components/nav/styles.ts
@@ -2,3 +2,4 @@ export const navAnimationVariants = {
 	open: { opacity: 1, x: 0 },
 	closed: { opacity: 0, x: '-100%' },
 }
+export const navItemAnimationStyle = 'transition-transform duration-200 active:scale-95'
